Let facilitators share answer in workshop/followup
When reaching the Workshop and Follow-up steps there should only exist
one answer from the facilitators. I.e. all facilitators will share and
edit the same answer for each question. One important caveat to note
is that behaviour is still undefined when multiple facilitators edit the
same question _at the same time_.

This commit implements this behaviour without breaking the backend API.
Instead it takes a shortcut inside the GetAnswer query where we dispatch
on role and progression. Hiding such behaviour might lead to
misunderstandings and confusion and we should strongly consider adding a
separate endpoint for shared answer updates.

In the backend, the author of a shared question will always be the
person that first created an answer. However, to avoid confusion for the
users the author is presented as 'Facilitators' in the frontend views.

As a direct consequence of this implementation two bugsi, both
introduced by [1] are also fixed:

1) Previously the users where presented with the _first_ answer to a
   question, regardless of whether it was their own answer or not.
2) Previously, when there existed more than one facilitators answer,
   only the _first_ answer where displayed in the AnswerSummarySidebar.

[1] SHA: ac154a882a40cc0d972f3c691f0375686af95d19