Introduce end-of-line normalization
The repository used a mix of dos (CRLF) and unix (LF) style end-of-line.
As most recommendations suggest, e.g. [1], this commit normalizes all
end-of-lines to unix. I.e. it removes the Carriage Return (CR).

No enforcements, i.e. in .gitattributes are added, and each contributor
is still responsible for configuring locally.

[1] https://docs.github.com/en/get-started/getting-started-with-git/configuring-git-to-handle-line-endings