Fix synchronization of AnswerSummarySidebar
This commit fixes a bug where changes to the QuestionList (users
changing filters), where not propagated correctly to the
AnswerSummarySidebar.

This entire commit is a hack. Which is a result of the QuestionList
being a javascript variable, while its filters are React states (hooks).
That means the update order is screwed up, and we have to do some manual
shenanigans. Additionally the fix is applied 3 different places as we
have duplicated implementations. Fixing these issues will require some
heavy refactoring, which is out-of-scope of this commit. Instead, it
adds to the mess...