Allow local run on custom ports
We are due to introduce some fake users for the testing. Unfortunately,
we still receive a certain amount of data from outside world which we
have no control of.

To be able to mock that data, we have to fake proper responses, thus
need to log in with a real users to know what to fake.

Additionally, it might be useful to be simultaneously logged in as
separate users during test development.

Cases above make it useful to be able to run bmt on custom ports.