Rename QuestionTemplate attr project -> projectId
It seams that a previous migration was runned from a dirty tree,
resulting in a missmatch between the database and the model in our code.
This resulted in errors when trying to apply new migrations.

We could add a new migration to sync up, however its easier to rename
our model.