Add DeleteAction mutation
Couple of notes:

1. Note has a required shadow Foreign Key and delete policy is set to
Cascade, which can be inferred from model defaults or checked in
Migration schemas.
That means that even as notes are not loaded when action is removed,
they are removed as well [1]

2. IQueryable can be useful in this case as due to Frontend
implementation we would need to know QuestionID to correctly remove
deleted action from cache.

To achieve that one option is to pass questionId from fronend. This
option has actually been used due to its simplicity.

Second option is to explicitly load related Question entry [2]. It leads
to a bit less clear design in our case, so this option has been ignored.

GraphQL seems to deal fine with returned IQueryables (GetAll), but in
that case returning IQueryable makes no sense as entry is deleted, so
IQueryable would be empty upon retrieval.

If Action is returned without any explicit load, Graphql doesn't deal
well with it and complains when attempting to retrieve non-nullable
field Question because it is not set.
Hence all the fields which are needed in frontend must be loaded
explicitly.

Approach also doesn't seem possible to be tested in unit tests
(Question is loaded by EF when asked for it) and we do not have graphql
tests implemented in frontend.

[1]
https://docs.microsoft.com/en-us/ef/core/saving/cascade-delete#required-relationship-with-dependentschildren-loaded

[2] https://docs.microsoft.com/en-us/ef/ef6/querying/related-data