Remove failing check for FollowUp - Actions
During development it has been missed entirely that on Follow Up step
Actions tab with all the actions is visible.

However due to randomization that has been somewhat caught, even though
incorrectly assumed to be a stability error.

In that case it makes sense to remove the existence-in-cache check from
tests as at the moment there is no strict requirement of place where
error is supposed to be displayed.