Use connection string for in-memory sqlite db
It looks like following happens:
- in certain places in the application several graphql requests are sent
almost at the same time
- second request is sent through the same connection before first one is
done processing
- sqlite can't deal with 2 pending requests on the same connection and
reports an error.

But it seems like we don't have to use the same connection in the
dbcontext and can use a connection string instead. It's probably then
that underlying framework deals with this for us and sends simultaneous
requests to sqlite in different connections.