Ignore unhandled promises
Most of instability that we are expecting in the future is related to
external fusion calls. Dealing with it from test to test appeared to be
tiresome.
As current situation showed, such calls can also lead to all test
failures where none are expected.

Thus the idea is to turn off cypress failures in such cases. We try to
do so only for the calls we already handle, and for fetching errors,
that sometimes appear instead.