Disconnect NoteCreateFrom from saving indicator
The saving indicator in ActionEditSidebar where responsible for the
status of both the Action mutations and the Note mutations in this
component. Having one indicator for to independent operation led to
several cases where it's status was either incorrect or ambiguous.

To disambiguate, the NoteCreateFrom is disconnected from the saving
state. The NoteCreateFrom has other visual clues to the current saving
state, such as disabling it's fields while saving, or only moving the
note from the input field to the note list when the save is complete.