Generalize Role based access in backend
The goal is to implement restrictions for which roles can do what
mutations in the backend. To that end the AssertIsFacilitator is
replaced by the more general AssertCanPerformMutation.

To avoid having to replicate AssertCanPerformMutation in the mocked
version of the AuthService, the method is also moved to the Mutation
class.