Add ProjectCategory table to Models
This commit adds a new Table "ProjectCategory" to the database model.

The ProjectCategory is part of the admin "workflow" and it represents a
questionaire set. The model will keep a single list of template question
- like before. A many-to-many relationship to QuestionTemplate is
created such that each QuestionTemplate can be reused across multiple
ProjectCategories. This layout ensures tracability of questions across
ProjectCategories while reducing the amount of items in the database.