Frontend change from delete to cancel action (#737)
Actions are now no longer deleted, they are cancelled (backend-term: voided). In the frontend we have replaced the trash icon with an X, and when the action is cancelled, we do not remove it, but show it greyed out and struck through. When cancelled, actions can no longer be edited, but they can still be viewed in the side panel.

Cancelled actions are no longer displayed in the action lists on the dashboard and on the follow up-step.