Progression-button only visible to users who can progress evaluation
Earlier, the progression-button was visible to everyone, even participants that can never progress the evaluation. Now, the button is only visible to the users who can progress the evaluation (such as facilitators). It will still be visible if it is disabled for these users, but will not be shown to users who will never click it.