Nomination: Erasing input after add (#802)
Problem: When nominating multiple users, after clicking "Add" on the first user, I have to manually erase the search query before typing the name of the next person I want to nominate. Often, this also results in the sidebar closing, because I accidentally click outside the view when highlighting the name I want to erase.

Solution: Now, when adding a user is successful, the search query and results are erased, so that I can more easily type in the next name I want to search for.