Evaluations can be hidden from nomination view (#819)
Facilitators and admins can now hide evaluations by clicking the "Hide"-button in Nomination view. Hidden evaluations are not visible in the normal evaluation views, but are visible on a "hidden"-tab which is only visible to admins.

Also added filter to user evaluations, so that hidden evaluations are not listed there.