Bug: All questions couldn't be scrolled into view (#884)
Problem: When scrolling in the questionnaire with the answers-sidebar open, the question opened in the sidebar is supposed to change when one question is scrolled out of view and another is scrolled into view. The problem was that sometimes the top or bottom question would not be changed in the sidebar.

Solution: The height of the scrolled element is now set dynamically so that it doesn't cause an additional scrollbar.

* Note that this solution might break if the tabslist in the views are removed.
* Also note that if you have a very large screen so that several questions are visible at the same time, you cannot scroll the bottom question into view because there is not enough space to scroll the currently selected question out of view. This can probably not be fixed within the current solution using scrolling.