Added url syntax to input field on closing comment (#885)
There has been a request that it should be possible to add a hyperlink when closing an action. A button for adding url with the correct syntax is now added close to the closing-reason input field.

Also changed the rendering of comments so that the link will be rendered as a link.