Remove force.click from unstable test
Test is unstable. Suspecting reason is the use of force: true, which
should be avoided.

Suspicion is that on fast CI dialog code sometimes is not yet removed
from the DOM when dialog is closed and when the test moves forwards.
Cypress uses click with force:true and clicks in the still existing text
of the dialog, instead of waiting for link to be clickable.