Removed outer scrollbar in questionnaire
To avoid the outer scrollbar, the inner div is set to a height that needs to be precisely the height of the view. This is calculated by fetching the heights and margins of some of the view elements, but for some reason the scroll bar got smaller, but didn't disappear. In this PR I simply added one pixel to the height, which made it finally disappear.